repo/classroom-server/requirements.txt
flask>=3.0.0
paho-mqtt>=1.6.0
amqtt>=0.11.0
